Fix duplicated gulpfiles in choose dialog
If you have multiple folders in your sublimem project and you've set the
`gulpfile_paths` setting set to `../` than it's possible to see the same
file multiple times in the choose dialog (one per folder).

To avoid this the path of each gulpfile gets normalized and only gets added to the
list if it's not in already.

Hopefully fixes #84 too.